requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
cryptography==43.0.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.2
django-cookies==0.0.2
django-cors-headers==4.4.0
django-countries==7.6.1
django-jazzmin==3.0.0
django-ninja==1.3.0
filelock==3.13.1
googletrans==3.0.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==24.1
pillow==10.4.0
platformdirs==4.2.0
psycopg==3.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.25.0
whitenoise==6.7.0
